+++ src/uet/oop/bomberman/entities/Tiles/Wall.java
@@ -5,7 +5,7 @@ import uet.oop.bomberman.entities.Entity;
 
 public class Wall extends Entity {
 
-    public Wall(float x, float y, Image img) {
+    public Wall(double x, double y, Image img) {
         super(x, y, img);
     }
